Вынес элементы экрана в отдельные файлы
